Implementado hacer pagos de pedidos pendientes

+++ app/Http/Controllers/PaymentController.php
@@ -8,6 +8,7 @@ use App\Enums\PaymentType;
 use App\Http\Helpers\Cart;
 use App\Models\CartItem;
 use App\Models\Order;
+use App\Models\OrderItem;
 use App\Models\Payment;
 use App\Models\Product;
 use Illuminate\Http\Request;
@@ -28,6 +29,7 @@ class PaymentController extends Controller
 
         //dd($cartItems);
         $total_price = 0;
+        $orderItems=[];
 
         foreach ($products as $product){
             $total_price+=$product->price * $cartItems[$product->id]['quantity'];
@@ -41,6 +43,11 @@ class PaymentController extends Controller
                     'unit_amount' => $product->price*100,
                 ],
                 'quantity' => $cartItems[$product->id]['quantity'],
+            ];
+            $orderItems [] = [
+                'product_id' => $product->id,
+                'quantity' => $cartItems[$product->id]['quantity'],
+                'unit_price' => $product->price
             ];
         }
 
@@ -75,6 +82,11 @@ class PaymentController extends Controller
             'session_id' => $checkout_session->id
         ];
 
+        foreach ($orderItems as $orderItem){
+            $orderItem['order_id']=$order->id;
+            OrderItem::create($orderItem);
+        }
+
         Payment::create($paymentData);
 
         return redirect($checkout_session->url);
@@ -82,6 +94,7 @@ class PaymentController extends Controller
 
     public function success(Request $request)
     {
+
         $user = $request->user();
 
         $stripe = new \Stripe\StripeClient(getenv('STRIPE_SECRET_KEY'));
@@ -99,9 +112,9 @@ class PaymentController extends Controller
 
             if(!$payment)
                 return view('payment.cancel',['message' => 'El pago no existe']);
-            else if ($payment->status!=PaymentStatus::Pending){
-                return view('payment.success');
-                //return view('payment.cancel',['message' => 'El pago ya se ha completado, gracias por su compra']);
+            else if ($payment->status!=PaymentStatus::Pending->value){
+                //return view('payment.success');
+                return view('payment.cancel',['message' => 'Este pago ya se  completado, gracias por su compra']);
             }
 
             $payment->status=PaymentStatus::Paid;
@@ -114,6 +127,8 @@ class PaymentController extends Controller
 
             $order->update();
 
+
+            if(!$_GET['resume'])
             CartItem::query()->where('user_id','=', $user->id)->delete();
 
             return view('payment.success' , compact('session' , 'customer'));
@@ -125,5 +140,44 @@ class PaymentController extends Controller
     public function cancel(Request $request)
     {
 
+    }
+
+    public function resumeOrderPayment(Request $request,Order $order)
+    {
+        $stripe = new \Stripe\StripeClient(getenv('STRIPE_SECRET_KEY'));
+
+        $orderItems = $order->orderItems;
+
+        $lineItems=[];
+
+        foreach ($orderItems as $orderItem){
+            $lineItems[] =[
+                'price_data' => [
+                    'currency' => 'eur',
+                    'product_data' => [
+                        'name' => $orderItem->product->title,
+                        'images' => [$orderItem->product->image],
+                    ],
+                    'unit_amount' => $orderItem->product->price*100,
+                ],
+                'quantity' => $orderItem->quantity,
+            ];
+        };
+
+        $checkout_session = $stripe->checkout->sessions->create([
+            'line_items' => $lineItems,
+            'mode' => 'payment',
+            'success_url' => route('payment.success',[],'yes').'?session_id={CHECKOUT_SESSION_ID}&resume=true',
+            'cancel_url' => route('payment.cancel',[],'yes'),
+            'customer_creation' => 'always'
+        ]);
+
+        $payment = $order->payment;
+
+        $payment->session_id = $checkout_session->id;
+
+        $payment->update();
+
+        return redirect($checkout_session->url);
     }
 }
